% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Timeline.R
\name{.MakeFlaggedGroupsTable}
\alias{.MakeFlaggedGroupsTable}
\title{Generate Flagged Groups Table}
\usage{
.MakeFlaggedGroupsTable(dfFlagged)
}
\arguments{
\item{dfFlagged}{Flagged timeline data frame.}
}
\value{
A data frame of flagged groups at the latest time point.
}
\description{
Internal function to create a table of currently flagged groups.
}
\keyword{internal}
