% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_TimeZFunnel.R
\name{Report_TimeZFunnel}
\alias{Report_TimeZFunnel}
\title{Generate a custom KRI HTML report}
\usage{
Report_TimeZFunnel(
  lCharts = NULL,
  dfResults = NULL,
  dfMetrics = NULL,
  dfGroups = NULL,
  strOutputDir = getwd(),
  strOutputFile = NULL,
  strInputPath = system.file("report", "Report_TimeZFunnel.Rmd", package = "gsm.timez")
)
}
\arguments{
\item{lCharts}{A list of charts to include in the report.}

\item{dfResults}{`data.frame` A stacked summary of analysis pipeline output.
Created by passing a list of results returned by [Summarize()] to
[BindResults()]. Expected columns: `GroupID`, `GroupLevel`, `Numerator`,
`Denominator`, `Metric`, `Score`, `Flag`, `MetricID`, `StudyID`,
`SnapshotDate`.}

\item{dfMetrics}{`data.frame` Metric-specific metadata for use in charts and
reporting. Created by passing an `lWorkflow` object to [MakeMetric()].
Expected columns: `File`, `MetricID`, `Group`, `Abbreviation`, `Metric`,
`Numerator`, `Denominator`, `Model`, `Score`, and `Threshold`. For more
details see the Data Model vignette: `vignette("DataModel", package =
"gsm.core")`.}

\item{dfGroups}{`data.frame` Group-level metadata dictionary. Created by
passing CTMS site and study data to [MakeLongMeta()]. Expected columns:
`GroupID`, `GroupLevel`, `Param`, `Value`.}

\item{strOutputDir}{The output directory path for the generated report. If not provided,
the report will be saved in the current working directory.}

\item{strOutputFile}{The output file name for the generated report. If not provided,
the report will be named based on the study ID, Group Level and Date.}

\item{strInputPath}{`string` or `fs_path` Path to the template `Rmd` file.}
}
\value{
Path to the generated HTML report (invisibly).
}
\description{
Wrapper around \code{\link[gsm.kri:Report_KRI]{gsm.kri::Report_KRI()}} that uses the custom report template
bundled with \code{gsm.timez} (\code{inst/report/Report_TimeZFunnel.Rmd}).
}
