% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag.R
\name{Flag}
\alias{Flag}
\title{Flag}
\usage{
Flag(
  dfAnalyzed,
  strColumn = "Score",
  vThreshold = c(-3, -2, 2, 3),
  vFlag = c(-2, -1, 0, 1, 2),
  vFlagOrder = c(2, -2, 1, -1, 0),
  vRiskScoreWeight = NULL,
  nAccrualThreshold = NULL,
  strAccrualMetric = NULL
)
}
\arguments{
\item{dfAnalyzed}{\code{data.frame} where flags should be added.}

\item{strColumn}{\code{character} Name of the column to use for thresholding. Default: \code{"Score"}}

\item{vThreshold}{\code{numeric} Vector of numeric values representing threshold values. Default is \code{c(-3,-2,2,3)} which is typical for z-scores.}

\item{vFlag}{\code{numeric} Vector of flag values. There must be one more item in Flag than thresholds - that is \code{length(vThreshold)+1 == length(vFlagValues)}. Default is \code{c(-2,-1,0,1,2)}, which is typical for z-scores.}

\item{vFlagOrder}{\code{numeric} Vector of ordered flag values. Output data.frame will be sorted based on flag column using the order provided. NULL (or values that don't match vFlag) will leave the data unsorted. Must have identical values to vFlag. Default is \code{c(2,-2,1,-1,0)} which puts largest z-score outliers first in the data set.}

\item{vRiskScoreWeight}{\code{numeric} Vector of weights to apply to each flag value. If provided, the output data.frame will have an additional column \code{Weight} with the corresponding weight for each flag. Default: \code{NULL}.}

\item{nAccrualThreshold}{\code{numeric} Specifies the minimum value required to return a \code{score} and calculate a \code{flag}. Default: NULL}

\item{strAccrualMetric}{\code{character} Specifies the Metric to apply \code{nAccrualThreshold} to in order to determine the validity of a flag. Options are "Numerator", "Denominator" or "Difference". If "Difference" is specified, the threshold is based on the difference between the Denominator and the Numerator for a given Group. Default: \code{NULL}.}
}
\value{
\code{data.frame} with an additional \code{Flag} column.
}
\description{
Alias for \code{gsm.core::Flag()}. Adds a Flag column to analyzed data identifying
possible statistical outliers based on threshold comparisons.
}
\seealso{
\code{\link[gsm.core:Flag]{gsm.core::Flag()}}
}
