% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.timez-package.R
\docType{package}
\name{gsm.timez-package}
\alias{gsm.timez}
\alias{gsm.timez-package}
\title{gsm.timez: gsm.timez - Timeline Z-Score Funnel Analysis GSM Extension}
\description{
A GSM extension for longitudinal site monitoring. Applies funnel plot scoring across sequential months to detect sites with unusual cumulative event trajectories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impala-consortium.github.io/gsm.timez}
}

}
\author{
\strong{Maintainer}: Donato Rolo \email{donato.rolo@roche.com}

Authors:
\itemize{
  \item Donato Rolo \email{donato.rolo@roche.com}
}

}
\keyword{internal}
