% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.timez-package.R
\docType{package}
\name{gsm.timez-package}
\alias{gsm.timez}
\alias{gsm.timez-package}
\title{gsm.timez: gsm.timez - Timeline Z-Score Funnel Analysis GSM Extension}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Donato Rolo \email{donato.rolo@roche.com}

}
\keyword{internal}
